backend/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "orion-backend"
version = "0.1.0"
description = "Orion notification gateway backend (FastAPI)"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "Orion Team"}
]
dependencies = [
  # Backend runtime
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.23.0",
  "pydantic-settings>=2.3.0",
  "SQLAlchemy>=2.0.0",
  "alembic>=1.13.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.25.0",
  # MySQL driver (pure Python)
  "pymysql>=1.0.3",
  # Auth
  "PyJWT>=2.8.0",
  "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
redis = ["redis>=5.0.0"]
celery = ["celery>=5.3.0"]
rq = ["rq>=1.16.0"]
apscheduler = ["APScheduler>=3.10.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
  "E", "F", "I",
  "B",
  "UP",
]
ignore = []
fix = true
